fixed leaks, updated def_ files, prototype piper.c complete
