+++ prototypes/util_str.h
@@ -3,8 +3,9 @@
 
 char **tokenize(char *text, char *delims, int limit);
 int str_nmatch(const char *txt_a, const char *txt_b, int n);
-void str_paste(char **dest, char *source);
-char *str_concat(char *pre, char *post);
-char *str_dup(char *text);
+void str_paste(char **dest, const char *source);
+char *str_concat(const char *pre, const char *post);
+char *str_dup(const char *text);
+int str_len(const char *text);
 
 #endif /* _CHEPE_STR_ */
